+++ packages/openmrs/test/utils.test.js
@@ -1,0 +1,128 @@
+import { expect } from 'chai';
+import { enableMockClient } from '@openfn/language-common/util';
+import testData from './fixtures.json' assert { type: 'json' };
+import { request } from '../src/Utils';
+
+const testServer = enableMockClient('https://fn.openmrs.org');
+const jsonHeaders = {
+  headers: {
+    'Content-Type': 'application/json',
+  },
+};
+const configuration = {
+  username: 'test',
+  password: 'strongpassword',
+  instanceUrl: 'https://fn.openmrs.org',
+};
+
+const state = {
+  configuration,
+  patient: testData.patient,
+  encounter: testData.encounter,
+};
+
+describe('request', () => {
+  it('should GET with a query', async () => {
+    testServer
+      .intercept({
+        path: '/ws/rest/v1/patient',
+        query: { q: 'Sarah 1' },
+        method: 'GET',
+      })
+      .reply(200, { results: testData.patientResults }, { ...jsonHeaders });
+
+    const { body } = await request(state, 'GET', '/ws/rest/v1/patient', {
+      query: { q: 'Sarah 1' },
+      baseUrl: state.configuration.instanceUrl,
+    });
+
+    expect(body.results[0].display).to.eql(testData.patientResults[0].display);
+  });
+  it('should auto-fetch patients with a limit', async () => {
+    testServer
+      .intercept({
+        path: '/ws/rest/v1/patient',
+        query: { q: 'Sarah', limit: 1 },
+        method: 'GET',
+      })
+      .reply(
+        200,
+        {
+          results: [{ display: 'Sarah 1' }],
+          links: [
+            {
+              rel: 'next',
+              uri: 'https://fn.openmrs.org/ws/rest/v1/patient?q=Sarah&limit=1&startIndex=1',
+              resourceAlias: null,
+            },
+          ],
+        },
+        { ...jsonHeaders }
+      );
+    testServer
+      .intercept({
+        path: '/ws/rest/v1/patient',
+        query: { q: 'Sarah', limit: 1, startIndex: 1 },
+        method: 'GET',
+      })
+      .reply(
+        200,
+        {
+          results: [{ display: 'Sarah 2' }],
+        },
+        { ...jsonHeaders }
+      );
+
+    const { body } = await request(state, 'GET', '/ws/rest/v1/patient', {
+      query: { q: 'Sarah', limit: 1 },
+      baseUrl: state.configuration.instanceUrl,
+    });
+
+    expect(body.results[0].display).to.eql('Sarah 1');
+    expect(body.results[1].display).to.eql('Sarah 2');
+  });
+
+  it('should not auto-fetch if the user sets startIndex', async () => {
+    testServer
+      .intercept({
+        path: '/ws/rest/v1/patient',
+        query: { q: 'Sarah', limit: 1 },
+        method: 'GET',
+      })
+      .reply(
+        200,
+        {
+          results: [{ display: 'Sarah 1' }],
+          links: [
+            {
+              rel: 'next',
+              uri: 'https://fn.openmrs.org/ws/rest/v1/patient?q=Sarah&limit=1&startIndex=1',
+              resourceAlias: null,
+            },
+          ],
+        },
+        { ...jsonHeaders }
+      );
+    testServer
+      .intercept({
+        path: '/ws/rest/v1/patient',
+        query: { q: 'Sarah', limit: 1, startIndex: 1 },
+        method: 'GET',
+      })
+      .reply(
+        200,
+        {
+          results: [{ display: 'Sarah 2' }],
+        },
+        { ...jsonHeaders }
+      );
+
+    const { body } = await request(state, 'GET', '/ws/rest/v1/patient', {
+      query: { q: 'Sarah', limit: 1, startIndex: 1 },
+      baseUrl: state.configuration.instanceUrl,
+    });
+
+    expect(body.results[0].display).to.eql('Sarah 2');
+    expect(body.results.length).to.eql(1);
+  });
+});
